requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
defusedxml==0.6.0
device==0.1
distlib==0.3.1
Django==3.1
django-appconf==1.0.4
django-ckeditor==6.0.0
django-froala-editor==3.2.1
django-imagekit==4.0.2
django-js-asset==1.2.2
djangorestframework==3.11.1
fake-useragent==0.1.11
filelock==3.0.12
idna==2.10
isodate==0.6.0
jdatetime==3.6.2
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
parse==1.16.0
pdfkit==0.6.1
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.5
pyaes==1.6.1
pyasn1==0.4.8
pyee==7.0.2
pygame==1.9.6
PyOpenGL==3.1.5
pyppeteer==0.2.2
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyquery==1.4.1
Pyrogram==0.18.0
PySocks==1.7.1
pysocks5==0.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-html==0.10.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
soap==0.0.1rc0
soupsieve==2.0.1
sqlparse==0.3.1
Telethon==1.16.2
TgCrypto==1.2.1
tqdm==4.48.2
urllib3==1.25.10
virtualenv==20.0.31
w3lib==1.22.0
websockets==8.1
zeep==3.4.0
